VERSION 1.0
LETS GOOOOO. Bullets now properly colored for players 1 and 2, they track to the top of the screen, if they intersect with an enemy the enemy takes damage and they are removed, and if they hit the top of the screen they are also removed. The basis for all future versions right here, a functional game.